requirements.txt
bokeh==0.12.16
chardet==3.0.4
click==6.7
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.0.1
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
packaging==17.1
pandas==0.22.0
Pillow==5.1.0
psutil==5.4.5
pyparsing==2.2.0
python-dateutil==2.7.2
python-Levenshtein==0.12.0
pytz==2018.4
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tornado==4.3
Werkzeug==0.14.1
xlrd==1.1.0